[user-progress] cleanup
