+++ frappe/public/build.json
@@ -61,7 +61,6 @@
 		"public/js/frappe/form/link_selector.js",
 		"public/js/frappe/form/multi_select_dialog.js",
 		"public/js/frappe/ui/dialog.js",
-
 		"public/js/frappe/form/controls/base_control.js",
 		"public/js/frappe/form/controls/base_input.js",
 		"public/js/frappe/form/controls/data.js",
@@ -164,6 +163,7 @@
 
 		"public/js/frappe/ui/page.html",
 		"public/js/frappe/ui/page.js",
+		"public/js/frappe/ui/slides.js",
 		"public/js/frappe/ui/find.js",
 		"public/js/frappe/ui/iconbar.js",
 		"public/js/frappe/form/layout.js",
@@ -171,8 +171,6 @@
 		"public/js/frappe/form/link_selector.js",
 		"public/js/frappe/form/multi_select_dialog.js",
 		"public/js/frappe/ui/dialog.js",
-		"public/js/frappe/form/user_progress_dialog.js",
-		"public/js/frappe/form/user_progress_slide.html",
 		"public/js/frappe/ui/app_icon.js",
 
 		"public/js/frappe/model/model.js",
